Use local git server for clones
Getting the history from a fetch is required to allow PBR to fetch the
project version and install. With '--depth 1' you'll see an install
that looks like 'keystone-0.0.0' which is bad.

This has a second purpose of allowing us to take advantage of
zuul-cloner. zuul-cloner does the proper merger of things for patches
that use Depends-On or are in a patch chain. Additionally, it uses the
local /opt/git git cache to futher reduce the fetching.

Change-Id: Ic0978dbb8e465b33c603d4151ec00922b206df48