Setup single script to all others
The original purpose for multiple scripts is long gone. Instead of
calling multiple scripts in the Dockerfile, we can call one that can
call others.

This also has the advantage of helping those that would provide thier
own SCRIPTS for CICD reasons of having to match only a single script.

Change-Id: Ib3e923e2e06660f7b18d16eb5e6f528c82c0e854