Add support for local registries
This patch does a few things. First, it gets rid of two variables and
uses only one variable now: WHEELS. No more DOCKER_REPO or DOCKER_TAG
env variables anymore.

If the path is not a url, it is assumed to be a docker image with or
without a tag. We then parse that into a registry, image, and tag. In
the case of no registry found, we assume DockerHub.

This currently does not support Auth on private registries.

Change-Id: I0d3b033dc5517622513b9028980edda8109baadc