Adjust the way the venv is built
It seems a newer package upstream that neutron depends on, depends on a
newer version of argparse to exist in the venv. Because of the way we
were creating the venv before, sometimes packages would not get
installed in the venv because they existed on the host. This was causing
problems for the pip requirements list build process. This patchset
relieves the issues found and hopes to prevent this type of breakage in
the future.

Change-Id: Ie52b8b8707ed5f9306adc78c7b50b1f28d634823