Only save compiled wheels
With this patch we stop trying to collect all possible packages we would
want and instead only save packages that we needed to build into wheels
ourselves because they were only distrubuted as source tarballs upstream

Doing it this way has a few advantages:
  * Reduced layer size to 50MB wheels from 250MB wheels
  * Can install additional packages that were not built into the wheels
image ahead of time (assuming they exist as wheels upstream)
  * Faster build times on slower WAN connections
  * Less I/O when building

We explicitly limit to binary installs with "--only-binary :all:" to
prevent any accidental source building.

This has no affect on building in an air-gapped environment. You would
already need a PyPI mirror to build in an air-gapped environment in the
first place.

Change-Id: I746d315ff642c951bbe798af7fbf951b5aa81a4f