Stop installing bindep via git
One of the claims we have made is that we can install in an airgapped
environemnt. This was true until we started installing master bindep to
take advantage of the new syntax.

Instead of installing master, we can just patch the file and remove the
patch when the new tag is created. Lets do that.

Change-Id: I7a86e6315c44e4e9aa050ce19865708dd6c90ff6