Make all the jobs voting by default
OpenStack test changes of main repos like nova, keystone, neutron
in the same pipeline.  These are meant to be delivered together,
and therefore I think the images that are built by loci should
be built together. It is currently in-line with current loci
code, as there is already many loci jobs like loci-keystone,
loci-nova, loci-neutron, in the main repository.

However, while all the loci-* jobs are passing, only the
loci-requirements is voting.

This is a problem, as code could (theoretically) merge while
breaking a test.

This should fix it.

Change-Id: I6173b57b2b2f79341a9c57acc1aa74f1a9d9facf