Install distro packages before installing project
This ensures the order is similar to the one in the
requirements project building [1].

It is also clearer that we install distro packages
that could be dependencies of the python code, before
installing the python code.

This has an impact, should people want to build
layers without using the requirements image build
process AND install packages only from sources and
not from wheels, due to auditing reasons.

[1]: https://github.com/openstack/loci/blob/06b5321eda7437c9a0bf150b8b6704385463a9e5/scripts/requirements.sh#L8-L9

Change-Id: I9f839d8dd74057a04c8fa8b5c3180c7a5b4864fb