Build OpenSUSE Leap 15 by default
This includes the building of OpenSUSE Leap 15 in checks/gates.

As the build of each project requires the building of the
requirements first, added to the fact that a new distro is added
into the jobs, this extends the timeout for the base job beyond
the maximum time for requirements, so that all the projects
(excluding requirements) have more time to complete.

Change-Id: I73b055b3942a6d80b8474544a9d65df0b3005e15