Updating binaries depenencies names for building suse images
For leap 15, some of the package names are different so updating their
names to reflect that. Verified these change when building openstack
service images from openstack helm. Openstack helm uses profile which
adds additional dependencies in images and may be that's why this issue
(different package name) is not encountered in loci internal build process.

Change-Id: Ic9dbdbe426f5d42c99a37bea0fc3137455bf5d5f