Fix error reporting in requirements.sh script
The 'echo %1' in requirements.sh did just that, echoed %1 without
reporting the source of the pip error. This patch changes this
command to 'cat $1', which dumps the content of the file referred
to by $1. This content is the pip package to install. Upon failure,
the user can know exactly which pip build failed and search for the
relevant error message in the build log.

Change-Id: Ieb5c928c31f1a6673152a0b1a15f6efe21cc063e