Fix broken symlinks from virtualenv 16.4.0 update
The virtualenv 16.4.0 updates fixes the handling of creating
symlinks of libraries to existing installations. Previously,
virtualenv would copy existing libraries into the virtualenv
making nested environments possible without retaining the
"parent" virtualenv. Now virtualenv builds a symlink chain,
which means we need to preserve the bootstrap virtualenv that
gives us an up-to-date installation of pip and virtualenv.

As distributions update their default virtualenv installations,
we might be able to use the --always-copy flag to stop using
symlinks and just copy dependencies over directly. However, the
virtualenv that ships with CentOS has a bug that causes the use
of that flag to fail, meaning as of right now we can't use it
to boostrap pip.

This patch moves the boostrap virtualenv from /tmp/venv, where
it is deleted in the build process, to /var/lib/pipboostrap,
where the symlink chain from

/var/lib/openstack -> /var/lib/pipbootstrap -> /usr/lib

will be remain unbroken.

Change-Id: I99124c2cfeb6ba7468a034ab510071eb67d98d66