Move CentOS to use RDO Stein packages
Pike is in extended maintenance and is EOLed in terms of RDO.

To be accurate the version of RDO repo should match with the branch
being used to build the containers. Anyway, IIUC, most python packages
are installed from source or pip, so i guess few packages should be
installed from distro repos.

Change-Id: I17354765c896fddc5962ee53d287d5a61abe8e11