python-ldap fix
In rocky and beyond pyldap becomes python-ldap. The structure of the
previous patch to fix them means we must always update the pydep.txt
list with a new release of openstack.

Reordering this we can explictly install only when the version is older
than rocky. Added openstack releases back to the Newton release for
compatibility.

Change-Id: Ibf88bd2113029d750f9a4f32e288e3d8be3a1bfd