OpenSuse: Cleanup script define packages to remove as string
This PS updates the SuSE cleanup script to define the virtual env
package to remove as a string, reather than as an array. An alternative
would have been to reference this via `${remove_packages[@]}` but as
this contains a singular value, and is not used anywhere else, it seems
simpler just to use a plain string here.

Change-Id: If559a710e4ec1f4ffc4528ecff58390c30a5351e
Signed-off-by: Pete Birley <pete@port.direct>