Add KEEP_ALL_WHEELS flag
The patch adds KEEP_ALL_WHEELS set to False by default.
The flag allows to do not remove wheels after built of
requirement image. This is useful where reproducibility
of builds with the same WHEEL image is needed as will allow
to keep 3rd party libs (not specified in upper-constraints.txt)
file with same versions.

Change-Id: I00fd3df2ba46072b3f19c9d08f69bb3c1f53d01f