Avoid storing cache in pip install for requirements image
If additional pip packages are provided for requirements image, they
will be installed into system before building wheels and, in addition,
wheel stored to local pip cache. Later, during wheels creation, this
cached wheel will be used and then deleted as system will think that
package is available in wheel form.

This only matters when package is not available in wheel form and needs
to be a part of requirements image.

Change-Id: I6d5410da02ff9717418de55876162d34b6185e1d