aarch64: enable building multi-arch images
At the moment, the only missing piece of being able to build
multiarchitecture images is the fact that loci does not support pulling
an image that has a 'fat manifest' (i.e. pointing at many different
images of different architectures but with the same name, built using
buildx).

This patch adds support to fetch-wheels in order for it to be able to
understand how these manifests work and download the correct one for the
architecture that is being built.

This has been tested locally here in our environments and successfully
produces multiarchitecture images using docker buildx.

Change-Id: I5ea132d12fec50b4c2cddd8b8c7c1f251ee2c30b