Reduce build targets
OpenSUSE LEAP is currently not in use by any known downstream
users and it has been broken for quite sometime, leaving the gate
in a broken state and unable to land any code.  There's no active
maintainer for it as well.

Debian support was added by us (VEXXHOST) but we decided that
we're not going to continue using it and move towards Ubuntu, so
we do not want to maintain it, as well as that it is building images
for Train only.

It also updates the README file accordingly as well as includes
focal in there which has been recently added.

Change-Id: Ideb497c169828184d301b6be4359a7c2228aa444