Specify a different source for novnc/spice-html5
The StarlingX project is seen some intermittence on the gitlab where
spice-html5 source is being cloned from and we would like to select a
different source. This changes gives the option to overwrite the source
repos and versions of both spice-html5 and novnc.

This will also provide the means to easily test new versions of those
libraries when in the event of an upversion.

Partial-Bug: #1956173
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I3001dad62a0caed26347dcb193d101c80901b5d8