Do not always install libapache2-mod-oauth2
This package is unavailable for 22.04 (the current LTS version of
Ubuntu). It is in 22.10 and may be backported to 22.04 through the
backports repo, but currently there is no prebuilt trusted source for
this package on 22.04.

This package is getting deployed on all ubuntu systems. In my opinion,
it should not have been merged, but we can gate the install around a
PROFILE='mod_oauth2' rather than revert the change entirely.

Related-Id: Ib3eac4269f7ac8365623f95a81b07a6436464ed0
Change-Id: I282b7b0aefa8aabc1f5c6495564d13b70cd0ec5d