Remove uwsgi by default
Not all deploys will use uwsgi by default but there is no way to disable
it currently. Since uwsgi itself isnt truly a python app (just a C
program with a python building shim).

The user will still be able to install binary package based uwsgi or
include uwsgi in the extra pip packages to install if they wish to.

Related-Id: I76612794c1ba8dbc45b62dff00cee43c6ba10a34
Change-Id: I15008d41633168fda31e061003ccf4681cade68e