Use treeless clone for project checkout
The openstack projects have fairly large repositories due to their long
history, which makes the checkout quite slow.

A treeless clone will get the whole history, but not the changes
themselves, which allows us to checkout any reference (branch or
commit-hash) directly, which will then fetch the files needed

For details see:
https://github.blog/2020-12-21-get-up-to-speed-with-partial-clone-and-shallow-clone/

The git repository is not added to the image. It is installed with pip and
resides as a standard package under
`/var/lib/openstack/lib/python<major>.<minor>/site-packages/`.
The only reason we probably need the history, is for the pip to determine
the version.