Add networking-baremetal to pydep.txt
Networking-baremetal package provides ironic-neutron-agent
and is not among the requirements of neither Neutron itself
nor global upper-constraints. So when we try to install it
using PIP_PACKAGES build env variable pip can not find it
among pre-downloaded wheels. So let's add it to pydep.txt
and when necessary to install it together with neutron
we can pass the baremetal profile.

Change-Id: I0f3eed5e1babd7fafcfbe2bae18a8b6b3b93397b