Build openstack-client image
The motivation behind is to have a consistent
openstack-client image related to a release or
with some changes being tested.

Loci builds so called `requirements` image that
contains a consistent set of packages related to
a particular release or from unreleased changes (for testing).
Then we install these packages from the `requirements`
image to all other images to have a consistent set
of images.

Change-Id: Ie8dac9fed8b09f5b6f2530c22ef9445d5c5cd462