Do not publish images on docker hub
The image promotion job often fails
due to Docker hub rate limits. The promotion
is run in the promote pipeline which starts
after the change is merged and we can not
use `recheck` to re-run the job. Let's
stop trying to publish images to Docker hub
and only publish them to quay.io (which is the
current default source of images).

Change-Id: Ife7ff860c465b3bbb20c437125b5e091619f67c0