Use python3 venv module instead of virtualenv
We install virtualenv module using apt package
python3-virtualenv. The Ubuntu Jammy version of
this package is not compatible with the Openstack 2025.1
(requirements/upper-constrains.txt:platformdirs===4.3.6).

Venv module is part of the standard Python3 distribution
since 3.3 and we don't need it to be compatible with Python2.
Let's use it instead of virtualenv.

Change-Id: Ic70af615ef42d69a4055998130487a5ed10d2232